indexing: update circulation policies mapping
* Adds analyzer on name and description fields.
* Fixes type on description field.
