+++ rero_ils/modules/circ_policies/mappings/v6/circ_policies/circ_policy-v0.0.1.json
@@ -17,13 +17,15 @@
         },
         "name": {
           "type": "text",
-          "copy_to": "circ_policy_name"
+          "copy_to": "circ_policy_name",
+          "analyzer": "global_lowercase_asciifolding"
         },
         "circ_policy_name": {
           "type": "keyword"
         },
         "description": {
-          "type": "keyword"
+          "type": "text",
+          "analyzer": "global_lowercase_asciifolding"
         },
         "organisation": {
           "properties": {
